bugfix nested iterable case
